refactor: create finish stricture of the project layout
