+++ src/components/app/App.ts
@@ -1,4 +1,4 @@
-import AppView from '../view/AppView';
+import AppView from '../view/appView';
 
 export default class App {
   view = new AppView();
