+++ src/components/view/layout/Footer.ts
@@ -1,0 +1,133 @@
+export default class Footer {
+  createElement() {
+    return `<footer class="footer">
+      <div class="footer__container container">
+        <p class="footer__copyright">© 2022 RSLang</p>
+        <ul class="footer__team-list">
+          <li class="footer__team-item">
+            <a class="footer__team-link" href="https://github.com/Axe11er" target="_blank">Alexander Naumov</a>
+          </li>
+          <li class="footer__team-item">
+            <a class="footer__team-link" href="https://github.com/alexborisewich" target="_blank">
+            Aliaksei Barysewich
+            </a>
+          </li>
+          <li class="footer__team-item">
+            <a class="footer__team-link" href="https://github.com/PiachkoAliaksey" target="_blank">Aliaksey Piachko</a>
+          </li>
+        </ul>
+        <a class="footer__rsschool-link" target="_blank" href="https://rs.school/js/">
+          <svg width="103" height="39" viewBox="0 0 103 39" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <g clip-path="url(#clip1)">
+              <path d="M53.1769 12.9178L58.0773 12.5949C58.189 13.4117 58.394 14.0196 58.7294 14.4565C59.2697
+              15.1404 60.015 15.4824 61.0025 15.4824C61.7292 15.4824 62.3068 15.3114 62.6981 14.9504C63.0707
+              14.6655 63.2943 14.2096 63.2943 13.7347C63.2943 13.2787 63.0894 12.8418 62.7354 12.5569C62.3627
+              12.2149 61.487 11.873 60.1082 11.5691C57.8537 11.0561 56.2326 10.3723 55.2824 9.49843C54.3321
+              8.68157 53.7918 7.48479 53.8291 6.21202C53.8291 5.33817 54.0899 4.46433 54.5744 3.74246C55.1333
+              2.92561 55.8973 2.27972 56.8102 1.91879C57.7978 1.48187 59.1766 1.25391 60.9094 1.25391C63.0335
+              1.25391 64.6545 1.65283 65.7724 2.46969C66.8904 3.26754 67.5612 4.55931 67.7661 6.326L62.9217
+              6.61094C62.7913 5.85108 62.5304 5.30018 62.1019 4.93924C61.6733 4.57831 61.1143 4.40734 60.3877
+              4.40734C59.7914 4.40734 59.3443 4.54032 59.0461 4.78727C58.7667 5.01523 58.5803 5.35717 58.599
+              5.7371C58.599 6.02205 58.748 6.288 58.9716 6.45897C59.2138 6.68693 59.7914 6.89589 60.7044
+              7.08586C62.9589 7.57977 64.5613 8.07368 65.5489 8.58659C66.5364 9.0995 67.2444 9.72639 67.673
+              10.4673C68.1201 11.2271 68.3437 12.101 68.3251 12.9938C68.3251 14.0576 68.027 15.1214 67.4307
+              16.0143C66.8159 16.9451 65.9588 17.667 64.9526 18.1039C63.8906 18.5788 62.549 18.8258 60.9466
+              18.8258C58.1145 18.8258 56.1581 18.2749 55.0588 17.1541C53.9595 16.0333 53.3074 14.6275 53.1769
+              12.9178ZM1.17384 18.5408V1.55785H9.76339C11.3471 1.55785 12.5769 1.69083 13.4153 1.97578C14.2352
+              2.24173 14.9618 2.79263 15.4463 3.5335C15.9866 4.35035 16.2475 5.30018 16.2288 6.288C16.2847 7.9597
+              15.4463 9.55542 14.0116 10.3913C13.4526 10.7142 12.8377 10.9422 12.2042 11.0561C12.67 11.1891 13.1359
+              11.4171 13.5458 11.683C13.8625 11.949 14.1234 12.2529 14.3656 12.5759C14.6451 12.8988 14.8873 13.2598
+              15.0923 13.6397L17.589 18.5598H11.7384L8.98083 13.3357C8.62681 12.6709 8.32869 12.2339 8.04921
+              12.025C7.67656 11.759 7.22938 11.626 6.7822 11.626H6.33502V18.5218H1.17384V18.5408ZM6.33502
+              8.43462H8.51502C8.98083 8.39662 9.428 8.32064 9.87518 8.20666C10.2106 8.14967 10.5087 7.9597
+              10.7136 7.67475C11.2167 6.99088 11.1422 6.02205 10.5273 5.43316C10.1919 5.14821 9.5398 4.99623
+              8.60818 4.99623H6.33502V8.43462ZM0 33.0922L4.90033 32.7692C5.01212 33.5861 5.21708 34.194 5.55246
+              34.6309C6.07417 35.3148 6.8381 35.6757 7.82562 35.6757C8.55228 35.6757 9.12989 35.5048 9.52117
+              35.1438C9.89381 34.8399 10.1174 34.403 10.1174 33.928C10.1174 33.4721 9.91245 33.0352 9.55843
+              32.7502C9.18578 32.4083 8.31006 32.0664 6.91263 31.7624C4.6581 31.2495 3.03708 30.5656 2.08683
+              29.6918C1.13658 28.8749 0.596237 27.6782 0.633502 26.4054C0.633502 25.5315 0.894356 24.6577 1.3788
+              23.9358C1.93777 23.119 2.7017 22.4731 3.61469 22.1122C4.60221 21.6752 5.98101 21.4473 7.71382
+              21.4473C9.83792 21.4473 11.4589 21.8462 12.5769 22.6631C13.6948 23.4799 14.347 24.7527 14.5519
+              26.5194L9.70749 26.8043C9.57706 26.0445 9.31621 25.4936 8.88766 25.1326C8.47775 24.7717 7.90015
+              24.6007 7.17348 24.6197C6.57724 24.6197 6.13007 24.7527 5.83195 25.0186C5.53383 25.2466 5.36614
+              25.5885 5.38477 25.9685C5.38477 26.2534 5.53383 26.5194 5.75742 26.6903C5.99964 26.9183 6.57724
+              27.1273 7.49023 27.3172C9.74476 27.8111 11.3471 28.305 12.3347 28.818C13.3222 29.3309 14.0302
+              29.9578 14.4588 30.6986C14.9059 31.4585 15.1295 32.3323 15.1295 33.2062C15.1295 34.27 14.8128
+              35.3148 14.2352 36.2076C13.6203 37.1385 12.7632 37.8603 11.7571 38.2973C10.695 38.7722 9.35347
+              39.0191 7.75109 39.0191C4.91896 39.0191 2.96255 38.4682 1.86324 37.3474C0.763929 36.2076 0.149059
+              34.8019 0 33.0922H0Z" fill="#323232"></path>
+              <path d="M24.7812 31.7624L29.2903 33.1492C29.0481 34.308 28.545 35.4098 27.8556 36.3786C27.2221
+              37.2334 26.3836 37.8983 25.4334 38.3352C24.4645 38.7722 23.2347 39.0001 21.7442 39.0001C19.9368
+              39.0001 18.4462 38.7342 17.3096 38.2023C16.1544 37.6704 15.1669 36.7205 14.3285 35.3718C13.49
+              34.023 13.0801 32.2943 13.0801 30.1857C13.0801 27.3742 13.8067 25.2086 15.2787 23.7079C16.7507
+              22.2071 18.8189 21.4473 21.5019 21.4473C23.6074 21.4473 25.247 21.8842 26.4582 22.739C27.6506
+              23.6129 28.545 24.9426 29.1412 26.7283L24.5949 27.7541C24.4831 27.3552 24.3154 26.9563 24.0918
+              26.6143C23.8124 26.2154 23.4583 25.9115 23.0298 25.6835C22.6012 25.4555 22.1168 25.3606 21.6324
+              25.3606C20.4585 25.3606 19.5642 25.8355 18.9493 26.8043C18.4835 27.5072 18.2413 28.628 18.2413
+              30.1477C18.2413 32.0284 18.5207 33.3201 19.0797 34.023C19.6387 34.7259 20.4213 35.0678 21.446
+              35.0678C22.4336 35.0678 23.1788 34.7829 23.6819 34.232C24.185 33.6431 24.5576 32.8262 24.7812
+              31.7624ZM35.3086
+              21.7322H40.4511V27.6781H46.0781V21.7322H51.2579V38.7152H46.0781V31.8384H40.4511V38.7152H35.3086V21.7322Z"
+              fill="#323232"></path>
+              <path d="M50.5498 30.2236C50.5498 27.4501 51.3137 25.2845 52.823 23.7458C54.3322 22.2071 56.4563
+              21.4282 59.158 21.4282C61.9342 21.4282 64.0583 22.1881 65.5675 23.7078C67.0768 25.2275 67.8221
+              27.3552 67.8221 30.0907C67.8221 32.0853 67.4867 33.7 66.8346 34.9728C66.201 36.2266 65.2135 37.2524
+              64.0024 37.9363C62.754 38.6391 61.2076 39.0001 59.3629 39.0001C57.4811 39.0001 55.9346 38.6961 54.7048
+              38.0882C53.4378 37.4423 52.3944 36.4355 51.705 35.2007C50.9411 33.871 50.5498 32.2183 50.5498
+              30.2236ZM55.6924 30.2426C55.6924 31.9523 56.0091 33.1871 56.624 33.947C57.2388 34.6878 58.0959
+              35.0678 59.1766 35.0678C60.2759 35.0678 61.133 34.7068 61.7479 33.966C62.3628 33.2251 62.6609
+              31.9143 62.6609 30.0147C62.6609 28.419 62.3441 27.2412 61.7106 26.5193C61.0771 25.7784 60.22
+              25.4175 59.1394 25.4175C58.1891 25.3795 57.2761 25.7974 56.6426 26.5383C56.0091 27.2792 55.6924
+              28.5139 55.6924 30.2426ZM73.095 30.2236C73.095 27.4501 73.859 25.2845 75.3682 23.7458C76.8774
+              22.2071 79.0015 21.4282 81.7032 21.4282C84.4795 21.4282 86.6222 22.1881 88.1128 23.7078C89.6034
+              25.2275 90.3673 27.3552 90.3673 30.0907C90.3673 32.0853 90.0319 33.7 89.3798 34.9728C88.7463 36.2266
+              87.7588 37.2524 86.5477 37.9363C85.2993 38.6391 83.7528 39.0001 81.9082 39.0001C80.0263 39.0001
+              78.4798 38.6961 77.2501 38.0882C75.9831 37.4423 74.9397 36.4355 74.2503 35.2007C73.4863 33.871 73.095
+              32.2183 73.095 30.2236ZM78.2376 30.2426C78.2376 31.9523 78.5543 33.1871 79.1692 33.947C79.7841 34.6878
+              80.6412 35.0678 81.7219 35.0678C82.8212 35.0678 83.6783 34.7068 84.2931 33.966C84.908 33.2251 85.2061
+              31.9143 85.2061 30.0147C85.2061 28.419 84.8894 27.2412 84.2559 26.5193C83.6224 25.7784 82.7653 25.4175
+              81.6846 25.4175C80.7343 25.3795 79.8027 25.7974 79.1878 26.5383C78.5543 27.2792 78.2376 28.5139 78.2376
+              30.2426Z" fill="#323232"></path>
+              <path d="M89.8272 21.7319H94.9698V34.5356H103V38.7149H89.8086V21.7319H89.8272Z" fill="#323232"></path>
+              <path d="M84.1338 20.7545C89.3069 16.7708 90.9038 10.0199 87.7005 5.6761C84.4972 1.33226 77.7068 1.04035
+              72.5337 5.02409C67.3606 9.00784 65.7638 15.7587 68.967 20.1025C72.1703 24.4464 78.9607 24.7383 84.1338
+              20.7545Z" stroke="#323232" stroke-width="0.589315" stroke-miterlimit="10"></path>
+              <path d="M84.1338 20.7545C89.3069 16.7708 90.9038 10.0199 87.7005 5.6761C84.4972 1.33226 77.7068 1.04035
+              72.5337 5.02409C67.3606 9.00784 65.7638 15.7587 68.967 20.1025C72.1703 24.4464 78.9607 24.7383 84.1338
+              20.7545Z" stroke="#323232" stroke-width="0.589315" stroke-miterlimit="10"></path>
+              <path d="M73.1141 11.6453L74.9773 10.3155L77.2691 13.6399C77.6604 14.1718 77.9585 14.7417 78.1821
+              15.3686C78.3125 15.8435 78.2752 16.3564 78.0889 16.8123C77.8467 17.3822 77.4554 17.8571 76.9337
+              18.1991C76.3188 18.636 75.7971 18.883 75.3499 18.9589C74.9214 19.0349 74.4742 18.9589 74.0643
+              18.769C73.6171 18.541 73.2259 18.2181 72.9277 17.8002L74.5301 16.2804C74.6606 16.4894 74.8282 16.6794
+              75.0146 16.8313C75.145 16.9263 75.2941 16.9833 75.4617 16.9833C75.5922 16.9833 75.7226 16.9263 75.8158
+              16.8503C76.0021 16.7363 76.1325 16.5084 76.1139 16.2804C76.058 15.9575 75.9276 15.6345 75.7226
+              15.3876L73.1141 11.6453ZM78.7038 13.7539L80.3993 12.3861C80.5857 12.6141 80.8279 12.7851 81.0887
+              12.88C81.4614 12.994 81.8527 12.918 82.1694 12.6711C82.393 12.5191 82.5793 12.3101 82.6539 12.0442C82.7657
+              11.7023 82.5793 11.3223 82.2439 11.2083C82.1881 11.1893 82.1322 11.1703 82.0763 11.1703C81.8527 11.1513
+              81.4614 11.2463 80.8838 11.4933C79.9335 11.8922 79.1882 12.0442 78.6293 11.9872C78.0889 11.9302 77.5858
+              11.6453 77.2877 11.1703C77.0641 10.8474 76.9523 10.4675 76.9337 10.0875C76.9337 9.65062 77.0455 9.2137
+              77.2877 8.85276C77.6417 8.33985 78.0703 7.90293 78.592 7.56099C79.3746 7.01009 80.064 6.74414 80.6602
+              6.74414C81.2564 6.74414 81.8154 7.02909 82.3371 7.61798L80.6602 8.96674C80.3248 8.52982 79.6913 8.43484
+              79.2628 8.77678L79.2069 8.83377C79.0205 8.94775 78.8901 9.11871 78.8156 9.32768C78.7597 9.47965 78.7969
+              9.65062 78.8901 9.7836C78.9646 9.87858 79.0764 9.95456 79.2069 9.95456C79.3559 9.97356 79.6168 9.89757
+              79.9894 9.72661C80.921 9.32768 81.6291 9.09972 82.1135 9.02373C82.5234 8.94775 82.952 8.98574 83.3433
+              9.15671C83.6973 9.30868 83.9954 9.57463 84.2004 9.89757C84.4612 10.2775 84.6103 10.7334 84.6289
+              11.2083C84.6475 11.7023 84.5171 12.1772 84.2563 12.5951C83.9209 13.108 83.4923 13.5449 82.9892
+              13.8869C81.9645 14.6087 81.126 14.9127 80.4552 14.7987C79.7286 14.6847 79.1323 14.3048 78.7038 13.7539Z"
+              fill="#323232"></path>
+            </g>
+            <defs>
+              <clipPath id="clip1">
+                <rect width="103" height="39" fill="white"></rect>
+              </clipPath>
+            </defs>
+          </svg>
+        </a>
+      </div>
+    </footer>`;
+  }
+
+  render(parent: HTMLBodyElement) {
+    parent.insertAdjacentHTML('beforeend', this.createElement());
+  }
+}
