+++ src/index.ts
@@ -1,4 +1,3 @@
-import 'normalize.css';
 import App from './components/app/App';
 import './sass/index.scss';
