+++ src/sass/abstracts/_constants.scss
@@ -1,1 +1,2 @@
-
+$error-color: #cc3333;
+$success-color: #4bb544;
