+++ src/sass/components/_timer.scss
@@ -1,16 +1,11 @@
 .timer {
-  &__svg {
-  }
+  &__svg {}
 
-  &__circle {
-  }
+  &__circle {}
 
-  &__path-elapsed {
-  }
+  &__path-expired {}
 
-  &__path-remaining {
-  }
+  &__path-remaining {}
 
-  &__counter {
-  }
+  &__counter {}
 }
